Initial Commit
Very Very Basic JavaFX application.

Foundation of the application is in the works still, this allows you to see the raw data, and the images, and allows you to 'overlay' the eye messages.

This overlay function is not properly implemented as it's a result of forgetting to clear the grid before adding in the new message display.

(a proper overlay option will be added eventually.